+++ core/src/lib/util.cc
@@ -2,7 +2,7 @@
    BAREOS® - Backup Archiving REcovery Open Sourced
 
    Copyright (C) 2000-2011 Free Software Foundation Europe e.V.
-   Copyright (C) 2016-2022 Bareos GmbH & Co. KG
+   Copyright (C) 2016-2024 Bareos GmbH & Co. KG
 
    This program is Free Software; you can redistribute it and/or
    modify it under the terms of version three of the GNU Affero General Public
@@ -37,6 +37,9 @@
 #include <sstream>
 #include <string>
 #include <vector>
+
+#include <openssl/rand.h>
+#include <openssl/err.h>
 
 // Various BAREOS Utility subroutines
 
@@ -319,11 +322,9 @@ char* encode_time(utime_t utime, char* buf)
   time_t time = utime;
 
 #if defined(HAVE_WIN32)
-  /*
-   * Avoid a seg fault in Microsoft's CRT localtime_r(),
+  /* Avoid a seg fault in Microsoft's CRT localtime_r(),
    * which incorrectly references a NULL returned from gmtime() if
-   * time is negative before or after the timezone adjustment.
-   */
+   * time is negative before or after the timezone adjustment. */
   struct tm* gtm;
 
   if ((gtm = gmtime(&time)) == NULL) { return buf; }
@@ -761,121 +762,32 @@ int DoShellExpansion(char* name, int name_len)
 }
 #endif
 
-/*
- * MAKESESSIONKEY  --  Generate session key with optional start
- *                     key.  If mode is TRUE, the key will be
- *                     translated to a string, otherwise it is
- *                     returned as 16 binary bytes.
- *
- *  from SpeakFreely by John Walker
- */
-void MakeSessionKey(char* key, char* seed, int mode)
+/* Create a new session key. key needs to be able to hold at least
+ * 120 bytes (keys are 40 bytes long, but errors might be longer).
+ * If successful, key contains the generated key, otherwise key will
+ * contain an error message. */
+bool MakeSessionKey(char key[120])
 {
-  int j, k;
-  MD5_CTX md5c;
-  unsigned char md5key[16], md5key1[16];
-  char s[1024];
-  constexpr int32_t ss = sizeof(s);
+  unsigned char s[16];
 
-  s[0] = 0;
-  if (seed != NULL) { bstrncat(s, seed, sizeof(s)); }
-
-  /*
-   * The following creates a seed for the session key generator
-   * based on a collection of volatile and environment-specific
-   * information unlikely to be vulnerable (as a whole) to an
-   * exhaustive search attack.  If one of these items isn't
-   * available on your machine, replace it with something
-   * equivalent or, if you like, just delete it.
-   */
-#if defined(HAVE_WIN32)
-  {
-    LARGE_INTEGER li;
-    DWORD length;
-    FILETIME ft;
-
-    Bsnprintf(s + strlen(s), ss, "%lu", (uint32_t)GetCurrentProcessId());
-    (void)!getcwd(s + strlen(s), 256);
-    Bsnprintf(s + strlen(s), ss, "%lu", (uint32_t)GetTickCount());
-    QueryPerformanceCounter(&li);
-    Bsnprintf(s + strlen(s), ss, "%lu", (uint32_t)li.LowPart);
-    GetSystemTimeAsFileTime(&ft);
-    Bsnprintf(s + strlen(s), ss, "%lu", (uint32_t)ft.dwLowDateTime);
-    Bsnprintf(s + strlen(s), ss, "%lu", (uint32_t)ft.dwHighDateTime);
-    length = 256;
-    GetComputerName(s + strlen(s), &length);
-    length = 256;
-    GetUserName(s + strlen(s), &length);
+  if (RAND_bytes(s, sizeof(s)) != 1) {
+    auto err = ERR_get_error();
+    ERR_error_string(err, key);
+    return false;
   }
-#else
-  Bsnprintf(s + strlen(s), ss, "%lu", (uint32_t)getpid());
-  Bsnprintf(s + strlen(s), ss, "%lu", (uint32_t)getppid());
-  (void)!getcwd(s + strlen(s), 256);
-  Bsnprintf(s + strlen(s), ss, "%lu", (uint32_t)clock());
-  Bsnprintf(s + strlen(s), ss, "%lu", (uint32_t)time(NULL));
-#  if defined(Solaris)
-  sysinfo(SI_HW_SERIAL, s + strlen(s), 12);
-#  endif
-  gethostname(s + strlen(s), 256);
-  Bsnprintf(s + strlen(s), ss, "%lu", (uint32_t)getuid());
-  Bsnprintf(s + strlen(s), ss, "%lu", (uint32_t)getgid());
-#endif
-  MD5_Init(&md5c);
-  MD5_Update(&md5c, (uint8_t*)s, strlen(s));
-  MD5_Final(md5key, &md5c);
-  Bsnprintf(s + strlen(s), ss, "%lu",
-            (uint32_t)((time(NULL) + 65121) ^ 0x375F));
-  MD5_Init(&md5c);
-  MD5_Update(&md5c, (uint8_t*)s, strlen(s));
-  MD5_Final(md5key1, &md5c);
-#define nextrand (md5key[j] ^ md5key1[j])
-  if (mode) {
-    for (j = k = 0; j < 16; j++) {
-      unsigned char rb = nextrand;
+
+  for (int j = 0; j < 16; j++) {
+    unsigned char rb = s[j];
 
 #define Rad16(x) ((x) + 'A')
-      key[k++] = Rad16((rb >> 4) & 0xF);
-      key[k++] = Rad16(rb & 0xF);
+    *key++ = Rad16((rb >> 4) & 0xF);
+    *key++ = Rad16(rb & 0xF);
 #undef Rad16
-      if (j & 1) { key[k++] = '-'; }
-    }
-    key[--k] = 0;
-  } else {
-    for (j = 0; j < 16; j++) { key[j] = nextrand; }
+    if (j & 1) { *key++ = '-'; }
   }
-}
-#undef nextrand
+  *--key = 0;
 
-void EncodeSessionKey(char* encode, char* session, char* key, int maxlen)
-{
-  int i;
-
-  for (i = 0; (i < maxlen - 1) && session[i]; i++) {
-    if (session[i] == '-') {
-      encode[i] = '-';
-    } else {
-      encode[i] = ((session[i] - 'A' + key[i]) & 0xF) + 'A';
-    }
-  }
-  encode[i] = 0;
-  Dmsg3(000, "Session=%s key=%s encode=%s\n", session, key, encode);
-}
-
-void DecodeSessionKey(char* decode, char* session, char* key, int maxlen)
-{
-  int i, x;
-
-  for (i = 0; (i < maxlen - 1) && session[i]; i++) {
-    if (session[i] == '-') {
-      decode[i] = '-';
-    } else {
-      x = (session[i] - 'A' - key[i]) & 0xF;
-      if (x < 0) { x += 16; }
-      decode[i] = x + 'A';
-    }
-  }
-  decode[i] = 0;
-  Dmsg3(000, "Session=%s key=%s decode=%s\n", session, key, decode);
+  return true;
 }
 
 /*