+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 - Disable automated package-tests for SLES 12 [PR #1677]
 - filed: fix vss during client initiated connections [PR #1668]
 - Improve FreeBSD dependencies [PR #1682]
+- stored: fix storage daemon crash if passive client is unreachable, create better session keys [PR #1702]
 
 ## [21.1.8] - 2023-10-23
 
@@ -716,4 +717,5 @@ and since Bareos version 20 this project adheres to [Semantic Versioning](https:
 [PR #1668]: https://github.com/bareos/bareos/pull/1668
 [PR #1677]: https://github.com/bareos/bareos/pull/1677
 [PR #1682]: https://github.com/bareos/bareos/pull/1682
+[PR #1702]: https://github.com/bareos/bareos/pull/1702
 [unreleased]: https://github.com/bareos/bareos/tree/master